+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendDocument.kt
@@ -78,6 +78,13 @@ data class SendDocumentData internal constructor(
     TextableSendMessageRequest<RawMessage>,
     ThumbedSendMessageRequest<RawMessage>
 {
+    init {
+        text ?.let {
+            if (it.length !in captionLength) {
+                throw IllegalArgumentException("Caption must be in $captionLength range")
+            }
+        }
+    }
 
     override fun method(): String = "sendDocument"
     override fun resultSerializer(): KSerializer<RawMessage> = RawMessage.serializer()