+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendVoice.kt
@@ -84,6 +84,13 @@ data class SendVoiceData internal constructor(
     ThumbedSendMessageRequest<RawMessage>,
     DuratedSendMessageRequest<RawMessage>
 {
+    init {
+        text ?.let {
+            if (it.length !in captionLength) {
+                throw IllegalArgumentException("Caption must be in $captionLength range")
+            }
+        }
+    }
 
     override fun method(): String = "sendVoice"
     override fun resultSerializer(): KSerializer<RawMessage> = RawMessage.serializer()