+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendAnimation.kt
@@ -95,6 +95,13 @@ data class SendAnimationData internal constructor(
     DuratedSendMessageRequest<RawMessage>,
     SizedSendMessageRequest<RawMessage>
 {
+    init {
+        text ?.let {
+            if (it.length !in captionLength) {
+                throw IllegalArgumentException("Caption must be in $captionLength range")
+            }
+        }
+    }
 
     override fun method(): String = "sendAnimation"
     override fun resultSerializer(): KSerializer<RawMessage> = RawMessage.serializer()